refactor: combine DocPageMeta with Meta component
Add DocPageMeta content in Meta.
Delete DocPageMeta.
