--- src/components/DocPageMeta/index.ts
@@ -1,2 +1,0 @@
-export {default as DocPageMeta} from './DocPageMeta';
-export * from './DocPageMeta';
